fixed bugs and misprints in invoking NST inference, made small improvements throughout the code

+++ bot/bot.py
@@ -3,6 +3,7 @@ import logging
 import os
 from urllib.parse import urljoin
 import aioboto3
+import aiobotocore
 import io
 import json
 
@@ -124,7 +125,7 @@ async def _get_userdata_for_model_outputs(id, assume_userdata_exist=False, incre
                result_counter = int(result_counter) if result_counter.isdigit() else None
             if increment_counter:
                 if result_counter is None:
-                    result_couner = 0
+                    result_counter = 0
                 result_counter += 1
                 await bucket.put_object(Key=f'{prefix}result_counter.txt', Body=str(result_counter).encode("utf-8"))
     if result_counter is not None:
@@ -164,7 +165,7 @@ async def _clean_userdata(id, assume_userdata_exist=False):
         if not is_userdata:
             await bucket.put_object(Key=prefix)
         async for obj in bucket.objects.filter(Prefix=prefix):
-            if obj.key != prefix:
+            if obj.key != prefix and obj.key != f'{prefix}result_counter.txt':
                 await obj.delete()
 
 
@@ -263,7 +264,7 @@ async def wrongcontent_handler(message: types.Message):
 async def _invoke_nst_endpoint(id, content_img_url, style_img_url):
     # noinspection PyBroadException
     try:
-        config = aiobotocore.config.Config(
+        config = aiobotocore.config.AioConfig(
             read_timeout=1200,
             retries={
                 'max_attempts': 0
@@ -272,40 +273,44 @@ async def _invoke_nst_endpoint(id, content_img_url, style_img_url):
         #sagemaker_runtime_client = boto3.client('sagemaker-runtime', config=config)
         #sagemaker_client = Session(sagemaker_runtime_client=sagemaker_runtime_client)        
         session = aiobotocore.get_session()
-        async with session.create_client('sagemaker-runtime', config=config) as sagemaker_client:
-            result_img_url = await _get_userdata_for_model_outputs(assume_userdata_exist=True, increment_counter=True)
-            _ = await client.invoke_endpoint(
+        async with session.create_client('sagemaker-runtime', **AWS_ACCESS_INFO_DICT, config=config) as sagemaker_client:
+            result_img_url = await _get_userdata_for_model_outputs(id, assume_userdata_exist=True, increment_counter=True)
+            sagemaker_response = await sagemaker_client.invoke_endpoint(
                 EndpointName='NeuralStyleTransfer',
                 Body=json.dumps({'content_url': content_img_url, 'style_url': style_img_url, 'result_url': result_img_url, 'max_epochs': 10, 'n_steps_in_epoch': 20}),
                 ContentType='application/json',
                 Accept='application/json'
             )
-        current_result_img_url = await _get_userdata_for_model_outputs(assume_userdata_exist=True, increment_counter=False)
+            response_body = sagemaker_response['Body']
+            async with response_body as stream:
+                data = await stream.read()
+                logging.info(json.loads(data.decode()))
+        current_result_img_url = await _get_userdata_for_model_outputs(id, assume_userdata_exist=True, increment_counter=False)
         if current_result_img_url == result_img_url:
             prefix = _get_prefix(id)
             async with aioboto3.resource("s3", **AWS_ACCESS_INFO_DICT) as s3:
-                bucket = await s3.Bucket(AWS_DEFAULT_BUCKET)
                 bucket_key = result_img_url[len("s3://") :]
                 bucket_name, key = bucket_key.split("/", 1)
-                result_s3_obj = await bucket.get_object(Key=key)
-                result_img_as_bytes = _read_s3_obj_as_bytes(result_s3_obj)
+                result_s3_obj = await s3.Object(bucket_name, key)
+                result_img_as_bytes = await _read_s3_obj_as_bytes(result_s3_obj)
             await bot.send_photo(id, result_img_as_bytes, caption='stylized via NST')                
     except Exception as e:
         logging.error(e)
-        await _clean_userdata(id)
-        await bot.send_message(id, f"Requests cannot be processed because the corresponding AWS resources are not launched, {MESSAGE_TO_CONTACT_SUPPORT}")
+        await bot.send_message(id, f"Request cannot be processed because the corresponding AWS resources are not launched, {MESSAGE_TO_CONTACT_SUPPORT}")
+    await _clean_userdata(id)
+    await bot.send_message(id, "For new request please send me your image with content to stylize")
 
 
 async def _invoke_cyclegan_endpoint(id, content_img_url, style_name):
     # noinspection PyBroadException
     try:
         # invoke_endpoint
-        await asyncio.sleep(45)
         raise Exception("Not yet implemented!")
     except Exception as e:
         logging.error(e)
-        await _clean_userdata(id)
-        await bot.send_message(id, f"Requests cannot be processed because the corresponding AWS resources are not launched, {MESSAGE_TO_CONTACT_SUPPORT}")
+        await bot.send_message(id, f"Request cannot be processed because the corresponding AWS resources are not launched, {MESSAGE_TO_CONTACT_SUPPORT}")
+    await _clean_userdata(id)
+    await bot.send_message(id, "For new request please send me your image with content to stylize")
 
 
 async def on_startup(dp):
